Create Improved_universal_dumper_extend.m
Modified version of the original universal dumper. Kept here until there's evidence that it works better than the original. It tries to overcome two weak points of the original script:
- The status message should now appear on all DryOS cameras, though good visibility is not guaranteed.
- It attempts to detect the firmware's actual address range and should now be compatible with DIGIC 7 cameras.
VxWorks cameras won't display any messages during dumping. Does not work on very early VxWorks models (Ixus 30/40, S1IS).

IMPORTANT: While this script attempts to display status messages on the screen, the text message may not be visible. If the camera appears not to respond to the script, check the card. If the script ran, status will be recorded in CBDUMPER.LOG and if the dump was successful, the ROM image will be saved in PRIMARY.BIN


Save this script as the file 'extend.m' on your card.